# removed the deep copy for the costmap_raw_ data using memcpy in prepareCostmap under costmap_2d_publisher.cpp