<!-- Please fill out the following pull request template for non-trivial changes to help us process your PR faster and more efficiently.-->

---

## Basic Info

| Info |  |
| ------ | ----------- |
| Ticket(s) this addresses   | #4914  |
| Primary OS tested on | Ubuntu |
| Robotic platform tested on | gazebo simulation |
| Does this PR contain AI generated software? | No |

---

## Description of contribution in a few bullet points

* I modified the loop that was being used to copy raw costmap by using a memcpy instead 

## Description of documentation updates required from your changes

* None would be required as it is just an enhancement 

---

## Future work that may be required in bullet points

* Further profiling costmap related functionalities to identify which processes are costing us the most.

#### For Maintainers: <!-- DO NOT EDIT OR REMOVE -->
- [ ] Check that any new parameters added are updated in docs.nav2.org
- [ ] Check that any significant change is added to the migration guide
- [ ] Check that any new features **OR** changes to existing behaviors are reflected in the tuning guide
- [ ] Check that any new functions have Doxygen added
- [ ] Check that any new features have test coverage
- [ ] Check that any new plugins is added to the plugins page
- [ ] If BT Node, Additionally: add to BT's XML index of nodes for groot, BT package's readme table, and BT library lists