@glitchhopcore 2025-02-13
Thanks for the review @SteveMacenski, I will try to do so and share some results here before and after using memcpy!