@glitchhopcore 2025-02-13
I have added some results from benchmarking the prepareCostmap() method over 10,000 iterations and its performance improvements before and after using memcpy.

For a local costmap of dimensions 10x10m:

Original method: ~287ms

![Untitled design](https://github.com/user-attachments/assets/4485a1e6-1d5f-4428-a468-84d9da0dc89b)

After memcpy implementation: ~24ms

![Untitled design(1)](https://github.com/user-attachments/assets/15fb8c1d-e815-4981-9d7d-1e8a8c42d51e)


For a global costmap of dimensions 100x100m:

Original method: ~33,888ms

![Untitled design(2)](https://github.com/user-attachments/assets/611ab003-6747-44de-a8a6-dd4b94e68253)

After memcpy implementation: ~5,607ms

![Untitled design(3)](https://github.com/user-attachments/assets/708b9c82-413b-48aa-8fb1-c292f3069271)

I used the following code for benchmarking these results:

```cpp
#include <chrono>
#include <iostream>
#include <random>
#include "rclcpp/rclcpp.hpp"
#include "rclcpp_lifecycle/lifecycle_node.hpp"
#include "nav2_costmap_2d/costmap_2d_publisher.hpp"
#include "nav2_costmap_2d/costmap_2d.hpp"

int main() {
    rclcpp::init(0, nullptr);
    auto node = std::make_shared<rclcpp_lifecycle::LifecycleNode>("costmap_bench");
    node->configure();
    node->activate();

    nav2_costmap_2d::Costmap2D cmap(2000, 2000, 0.05, 0.0, 0.0); // or 200, 200, 0.05 for local costmap
    
    std::mt19937 gen(std::random_device{}());
    std::uniform_int_distribution<> dist(0, 255);
    
    for(unsigned int i = 0; i < cmap.getSizeInCellsX(); i++) {
        for(unsigned int j = 0; j < cmap.getSizeInCellsY(); j++) {
            cmap.setCost(i, j, dist(gen));
        }
    }

    nav2_costmap_2d::Costmap2DPublisher pub(node, &cmap, "map", "costmap", true, 0.0);

    auto t1 = std::chrono::high_resolution_clock::now();
    for(int i = 0; i < 10000; i++) {
        pub.prepareCostmap();
    }
    auto t2 = std::chrono::high_resolution_clock::now();
    
    std::cout << "Time taken with memcpy: " << std::chrono::duration_cast<std::chrono::milliseconds>(t2 - t1).count() << "ms\n";

    node->deactivate();
    node->cleanup();
    node->shutdown();
    rclcpp::spin_some(node->get_node_base_interface());
    rclcpp::shutdown();
    return 0;
}