@SteveMacenski 2025-02-13
Ah very nice and perfect, a good 5-10x speed up!